requirements.txt
transformers[sentencepiece]
datasets
tqdm
pandas
pyarrow
fastparquet
pandavro
avro
google-cloud-aiplatform
scikit-learn
